+++ mip/regs.c
@@ -2384,7 +2384,7 @@ temparg(struct interpass *ipole, REGW *w)
 			continue; /* arg in register */
 		if (w != &nblock[regno(p->n_left)])
 			continue;
-		w->r_color = p->n_right->n_lval;
+		w->r_color = (int)p->n_right->n_lval;
 		tfree(p);
 		/* Cannot DLIST_REMOVE here, would break basic blocks */
 		/* Make it a nothing instead */
